requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Async support  
aiofiles>=0.8.0

# LLM providers (REQUIRED - at least one)
openai>=1.0.0
anthropic>=0.7.0

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20

# Utilities
requests>=2.28.0
gitpython>=3.1.0
psutil>=5.9.0